+++ docs/RELEASES.md
@@ -17,7 +17,7 @@ Legend:
 ## <a name="7.2.0.0"></a>7.2.0.0 :whale: <sup>`5`</sup>
 - [ ] Fix moving tasks from TODO to RELEASE
 - [ ] Fix search with diacritics [#107]
-- [ ] `210815>......` Feature to unify Categories
+- [x] `210815>210717` Feature to unify Categories
 - [x] `210815>210816` Hide warning about being inactive in the first three days
 - [x] `210814>210815` Add edit move button to search result